Opprydding i skif storetest tester - Tatt bort ws skjema for "demo" og lagt til ws skjema for "basic"
[git-p4: depot-paths = "//skif/SkifKode/trunk/": change = 49392]

+++ skif-store-test/skif-store-test-api/src/main/java/no/statkart/skif/storetest/domain/basic/HistSimple.java
@@ -11,7 +11,7 @@ import no.statkart.skif.storetest.domain.AbstractStoreTestBubbleWithHistory;
 public class HistSimple extends AbstractStoreTestBubbleWithHistory{
     private static final long serialVersionUID = 1L;
 
-    private long nr;
+    private int nr;
     private String text;
     private int testSetNumber;
 
@@ -27,11 +27,11 @@ public class HistSimple extends AbstractStoreTestBubbleWithHistory{
         return (HistSimpleId<?>) super.getId();
     }
 
-    public long getNr() {
+    public int getNr() {
         return nr;
     }
 
-    public void setNr(long nr) {
+    public void setNr(int nr) {
         this.nr = nr;
     }
